PoC: use miqt instead of DOtherSide
`DOtherSide` is a hand-made built C++ -> C wrapper that exposes a small
subset of Qt.

With an approach like
https://github.com/arnetheduck/nim-miqt-poc/pull/1, the full qt api,
including both qml and regular Qt, can be mapped to C allowing access to
all of qt in a low-maintenance manner, simplifying upgrades as well.

At the same time, the way the poc is written, the C++->C wrappers are
generated class by class and therefore don't need a separate compilation
step to first compile `DOtherSide` to a library which then is linked to
- instead, `{.compile.}` takes care of compiling the correct minimal set
of qt using the right flags from `pkg-config`. There is also no need to
write "nim headers" like `dotherside.nim` that import the c code to nim
since these trivially can be generated as well.

This PR contains a poc-quality compatibility layer for qml that instead
of dotherside imports miqt-poc. It's a bit janky of course because of
how miqt-poc itself is just an experiment, but it shows the approach
more or less - a more complete approach would remove the `DosQ*`-style
classes as well.

With this approach, it also becomes possible to share the same c backend
between `nimqml` and `nimqt`, reducing the overall maintenance burden of
bringing Qt to Nim - the C++->C mapping step can also potentially be
shared between other languages (go/zig/etc) which further speaks for the
approach (either from a high-level "share-notes" level or by actually
using the same C bindings).

The PoC is enough to run the `helloworld` example assuming the import
paths in `nim.cfg` are updated, and that
https://github.com/arnetheduck/nim-miqt-poc/pull/1 (revision
https://github.com/arnetheduck/nim-miqt-poc/commit/c87d64243c66e9e57b41c5641a9ba8f8ec0e59d6
at the time of writing) is used